requirements.txt
dotenv


# Required for db and data massaging
google-cloud-firestore
pandas
requests
